refactor(listen): split upcoming rows
